updated cleanRun.sh

+++ cleanRun.sh
@@ -1,4 +1,12 @@
+#!/bin/bash
 
-run=$1
+if [ $# -eq 1 ]; then
 
-rm -rf /fff/BU0/ramdisk/run${run} /fff/BU0/data/sm/run${run} /fff/data/run${run} /fff/BU0/output/run${run}
+  run=$1
+  rm -rf /fff/BU0/ramdisk/run${run} /fff/BU0/data/sm/run${run} /fff/data/run${run} /fff/BU0/output/run${run}
+
+else
+
+  printf "\n%s\n\n" "[cleanRun.sh] ERROR: one command-line argument required -> specify run number (example: ./cleanRun.sh [RUN])"
+  exit
+fi
